Added some nasty Java hacks, few things need to be figured out first though
I added some Java reflection hacks to get a few pieces of the core of
spigot so I can listen to all events (required for the warn command and
future freeze command) but java has an illegal argument exception about
a final class. I will need to look into this further but for now I'll put
this code here.