Changed the runnable to Future Task so we could properly do exception handling, also added some translation util for doing configuration files.
The translation util is still very buggy and doesn't currently work. But
this will allow us to things like {PLAYER} and parsing it easily. Along
with this, futuretask should allow us to handle the sql exceptions
better than using local booleans in the databaseutils class and setting
it during the runnable.