Fixed TranslationUtil, Moved some SQL queries to be async
TranslationUtil now works as it should, Java doesn't have a convenient
way to check if a character is a hexidecimal so I had to write my own.
The utility now also correctly parses variables out of a message and
replaces them so the messages we use can be configurable.

Spigot's BukkitScheduler is absolute garbage for being incapable of
accepting a FutureTask and not freezing the server thread. Now I used
a thread pool spawned at the launch of the plugin to execute all async
queries to the database so we load and execute faster.