Settled on a library to handle IP addresses
Picked a library which seems to do most IP address comparison
functions we'll need for this plugin. Rewrote much of the
ConnectionListener.java class to now kick people in the following order:

- UUID Bans
- IP Bans
- Pending unaccepted warnings

Updated the TODO with a few more things on IP banning but tomorrow
I can start on the /ipban command at least and start tracking ips of
when players are banned.