Updated SQL tables to store player IP addresses.
Store the last known IP address of a player in the database so the web
frontend has something to display.

Changed how the Messages class worked a bit so it will now throw an
exception if the message node in the yaml file does not exist or cannot
be found. This is an error for the user to ensure they configured things
right. Right now, many things are throwing this exception and the config
needs to be updated to reflect the missing strings. This should probably
be handled better in the future but that's a tomorrow problem.