[Commands] Fix providing the -s argument (untested)
Removed the old code for handling the -s argument to most commands
for silencing them (not announcing them publically). This is the
correct way to handle arguments to commands but requires some adjusting
of the code. I suspect -s will become part of the reason message so that
will need to be fixed.

Fixed some of the messages for some of the commands to use the messages.yml
file instead of hard-coded messages.