[Permissions] Revamped how permissions worked a bit
This still needs a ton of work but at least there is a utility
that can check permissions for almost anything given to it.
This may later need to be fixed up for more complex permission
checks and I would like to add a Java annotation maybe in the
future which can make things easier if done right.