[IP Bans] Check for alt accounts on connect
Check to see if the user connecting with the same IP address
is an alt account or not. At the moment this just announces to
all the server OPs but it might be desirable to kick/ban the
alt account if they're already banned.

Added a new BroadcastUtil class which can help with broadcasting to
op groups for those who are tuned into them. I may extend this to do
additional announcements of different sorts.

@Zachery will have to test this as I don't have alt accounts.