Added broadcast and report command
Both of these commands have been added but are untested, they should
work.

FindPlayerByBanID should be async now and no longer running on the main
thread, hopefully the server wont hang when a new user is banned.

I also moved the history command to it's own folder and added
StaffHistoryCommand with it.