Discord simplified messages now work and fixed bugs
The Discord simplified messages returned a weird java string because we
were trying to add an arguemnt to it (ex. "%s message", args).
Java didn't like that, so I removed it and replaced it with the new
method of adding config messages that we created a while ago, i'm not
sure why this wasn't updated before.

A couple bugs have been fixed, the fixed bugs are as following;

  The database queries wern't changed in ConnectionListener, so it would
  kick you from the server.

  A couple of the commands wern't updated to the new config parser
  resulting in NPEs everywhere! (AAAH)

  The unmute command was returning a message to console saying it was
  unmuted, this wasn't intended and I have made it so it only sends it to
  the unmuted player if they're online