Rewrote the /history command
History now shows a list of stuff instead of that condensed mess we had before.
The look of it can be configured easily in messages.yml and it's quite flexible.
I have also added pagination which lets lots of things be displayed in a set of
nice pages with a configurable pagination line. I may make the number of pages
configurable too but we'll see cuz that could be too confusing.

I also fixed a bug in the color parser. Minecraft is not just hexidecimal color
codes but it also has r, k, l, m, n, o, p color codes as well which were previously
not interpreted by the color parser.