Added a new Punishment object for managing punishments better.
The new object will allow for better management of the punishment table
in the database. This object further cleans up the codebase and makes
things much more readable. This also puts the task for adding/removing
punishments to the object itself and not DatabaseUtil.

Paginator for /history and /staffhistory can now have it's page size
configured from messages.yml making pages easier to work with.

Added option to kick players that have been previously banned as an
alternative account.

Added "Website" as a global variable in messages.yml