Added /prunehistory and a new Command object
The new prunehistory command is done but untested and probably a huge mess
much of the command is kinda hacked together but will be cleaned up when
I have a more clear mind to know what I am doing.

Renamed `Executioner` to `Arbiter` since that is a more applicable name for who
that person is (the person who ran the command to ban someone). I also
renamed `AppealStaff` to `Appellee` according to US legal glossery.

Added a new RistExCommand class which will handle syntax errors better than
the default syntax error handler which is part of spigot's trashy plugin.yml
bullshit way of handling syntax errors. This will also allow us to make use
of dynamic syntax errors and use of messages.yml so it can be colorized/customized
and translated in the future.