Fuck entirely off with Spigot's CommandExecutors and plugin.yml, also added /unwarn.
This is a huge rewrite of how commands are processed both in spigot and in lolbans.
Instead of registering each CommandExecutor and doing some knobheaded way of handling
commands with plugin.yml, I directly created each command itself and just dumped it
into the CommandMap using reflection (because knobhead didn't add getCommandMap() as
a public API in the Server class but CommandMap is entirely documented). It seems like
the more and more I have to use Spigot/Bukkit APIs, the more and more I have to make
some hacky way to do things the way they should've been done in the first place.
This also makes command handling much more understandable and removes all the knobbery
in plugin.yml. Ideally I would like to get plugin.yml to be nearly empty but I need to
make permissions go away next.