Bug fixes and minor features
The "CENSOREDIPADDRESS" config node was added for people who do not want
to display IPs to the public.

The Punishment object now uses the senders UUID instead of the
CommandSender, this makes using it outside of a command easier.

Offline bans can track IPs as long as users have joined before, this is
the same for username based IP bans.